ensure all input output only over input_output module
